+++ parser/CMakeLists.txt
@@ -10,6 +10,16 @@ add_library(scarParser STATIC
     parse_binop.cc
     parse_unop.cc
     parse_constants.cc
+
+    semantic_analysis/analyze_block.cc
+    semantic_analysis/analyze_declaration.cc
+    semantic_analysis/analyze_expression.cc
+    semantic_analysis/analyze_statement.cc
+
+    pretty_print/pretty_print_block.cc
+    pretty_print/pretty_print_declaration.cc
+    pretty_print/pretty_print_expression.cc
+    pretty_print/pretty_print_statement.cc
 )
 
 set_basic_compile_options(scarParser)
